pyproject.toml
[project]
name = "veragent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "CC-BY-NC-SA-4.0"}
dependencies = [
    # 核心依赖
    "openai>=1.0.0,<2.0.0",
    "requests>=2.25.0,<3.0.0",
    "python-dotenv>=0.19.0,<2.0.0",
    "pydantic>=2.0.0,<3.0.0",

    # 基础工具库
    "beautifulsoup4>=4.9.0,<5.0.0",
    "numpy>=2.0.0,<3.0.0",
    "networkx>=2.6.0,<4.0.0",
    "tiktoken>=0.5.0",

]

[tool.uv.sources] # 如果使用 uv
veragent = { path = "./ver_agent", editable = true }